Fix timeout underflow when framerate < fps cap
When I left that comment saying "// lotta casting, oof" turns out that
was indeed too much casting, as when I casted to usize back to i32 it
would underflow if the delay was more than the FPS cap. I planned to do
this initially hence the `.max(0)` but I just started blindly going
through complie errors adding casts and kinda forgot.

+++ src/main.rs
@@ -196,8 +196,7 @@ fn main() {
                     let now = std::time::Instant::now();
                     matrix = gol_step(&matrix, live, birth);
                     redraw_all!();
-                    // lotta casting, oof.
-                    window.timeout((max_delay as u128 - now.elapsed().as_millis()).max(0) as i32);
+                    window.timeout((max_delay - now.elapsed().as_millis() as i32).max(0));
                 }
                 window.timeout(-1);
                 pancurses::curs_set(1);
